Version 9.0.4.
Chris sent me this tarball today, but it dates from 2012.  The CHANGELOG is
short ("Fixed yaap measure (Stephen Robertson)") but there seem to be a
lot of changes in a lot of files from 9.0a3, including several measure files.
This commit just brings in his code without even compile checking or running tests.